requirements.txt
python-telegram-bot==21.6
anthropic>=0.34.0
gspread>=6.1.2
google-auth>=2.29.0
apscheduler>=3.10.4
python-dotenv>=1.0.1
